+++ app/src/main/java/com/example/android_accountbook_13/presenter/component/AccountBookHistoryItem.kt
@@ -77,8 +77,8 @@ fun AccountBookHistoryItemContent(
 @Composable
 fun AccountBookHistoryItemHeader(
     date: String,
-    income: Long,
-    expense: Long,
+    income: Int,
+    expense: Int,
     leftChecked: Boolean,
     rightChecked: Boolean,
 ) {
